requirements.txt
streamlit>=1.35.0
google-generativeai>=0.7.0
requests>=2.31.0
PyJWT>=2.8.0
supabase>=2.4.0
